requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
requests==2.31.0
